CRQ-2810 Added fix to handle force expands on Functions into JSON writer and refactored
